chore: remove klass.patch so it doesn't get used.
The sole usage of this is within snakeoil, and tooling
that monkeypatches module namespaces really isn't something
even *snakeoil*- despite the name- should enable.

Besides, there's far better external libraries for this.
